# web-ui/openapi.json
{
  "components": {
    "schemas": {
...
      "CheckpointSummary": {
        "additionalProperties": false,
        "properties": {
          "compatible": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compatible"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "exists": {
            "title": "Exists",
            "type": "boolean"
          },
          "pipeline_fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pipeline Fingerprint"
          },
          "pipeline_name": {
            "title": "Pipeline Name",
            "type": "string"
          },
          "sequence": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sequence"
          },
          "source_cursor": {
            "$ref": "#/components/schemas/JSONValue-Output"
          },
          "state_nodes": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "State Nodes",
            "type": "array"
          }
        },
        "required": [
          "pipeline_name",
          "exists"
        ],
        "title": "CheckpointSummary",
        "type": "object"
      },
...
      "DataSourceConfig-Input": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/JSONValue-Input"
          },
          "format": {
            "$ref": "#/components/schemas/InputFormat"
          },
          "id": {
            "pattern": "^[A-Za-z][A-Za-z0-9_-]{0,63}$",
            "title": "Id",
            "type": "string"
          },
          "input_name": {
            "default": "input",
            "maxLength": 160,
            "minLength": 1,
            "title": "Input Name",
            "type": "string"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          }
        },
        "required": [
          "id",
          "format",
          "data"
        ],
        "title": "DataSourceConfig",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProjectConfig-Input": {
        "additionalProperties": false,
        "properties": {
          "data_sources": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/DataSourceConfig-Input"
            },
            "title": "Data Sources",
            "type": "array"
          },
          "description": {
            "default": "",
            "maxLength": 2000,
            "title": "Description",
            "type": "string"
          },
          "format_version": {
            "const": "1",
            "default": "1",
            "title": "Format Version",
            "type": "string"
          },
          "id": {
            "pattern": "^[A-Za-z][A-Za-z0-9_-]{0,63}$",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "pipeline": {
            "$ref": "#/components/schemas/PipelineConfig"
          },
          "run_options": {
            "$ref": "#/components/schemas/RunOptions",
            "default": {
              "max_input_bytes": 10485760,
              "max_rows": 100000,
              "memory_limit_mb": 512,
              "output_rows": 1000,
              "timeout_seconds": 30.0
            }
          }
        },
        "required": [
          "id",
          "name",
          "pipeline"
        ],
        "title": "ProjectConfig",
        "type": "object"
      },
      "ProjectConfig-Output": {
        "additionalProperties": false,
        "properties": {
          "data_sources": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/DataSourceConfig-Output"
            },
            "title": "Data Sources",
            "type": "array"
          },
          "description": {
            "default": "",
            "maxLength": 2000,
            "title": "Description",
            "type": "string"
          },
          "format_version": {
            "const": "1",
            "default": "1",
            "title": "Format Version",
            "type": "string"
          },
          "id": {
            "pattern": "^[A-Za-z][A-Za-z0-9_-]{0,63}$",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "pipeline": {
            "$ref": "#/components/schemas/PipelineConfig"
          },
          "run_options": {
            "$ref": "#/components/schemas/RunOptions",
            "default": {
              "max_input_bytes": 10485760,
              "max_rows": 100000,
              "memory_limit_mb": 512,
              "output_rows": 1000,
              "timeout_seconds": 30.0
            }
          }
        },
        "required": [
          "id",
          "name",
          "pipeline"
        ],
        "title": "ProjectConfig",
        "type": "object"
      },
      "ProjectSummary": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "node_count": {
            "title": "Node Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "node_count"
        ],
        "title": "ProjectSummary",
        "type": "object"
      },
...
      "RunRequest": {
        "additionalProperties": false,
        "properties": {
          "inputs": {
            "additionalProperties": {
              "$ref": "#/components/schemas/InputPayload"
            },
            "title": "Inputs",
            "type": "object"
          },
          "options": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunOptions"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "RunRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Calc Flow API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "operationId": "list_projects_api_v1_projects_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectSummary"
                  },
                  "title": "Response List Projects Api V1 Projects Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Projects"
      },
      "post": {
        "operationId": "create_project_api_v1_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectConfig-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectConfig-Output"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Project"
      }
    },
    "/api/v1/projects/import": {
      "post": {
        "operationId": "import_project_api_v1_projects_import_post",
        "parameters": [
          {
            "in": "query",
            "name": "format",
            "required": true,
            "schema": {
              "pattern": "^(json|yaml)$",
              "title": "Format",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "replace",
            "required": false,
            "schema": {
              "default": false,
              "title": "Replace",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectConfig-Output"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Project"
      }
    },
    "/api/v1/projects/{project_id}": {
      "delete": {
        "operationId": "delete_project_api_v1_projects__project_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Project"
      },
      "get": {
        "operationId": "get_project_api_v1_projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectConfig-Output"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project"
      },
      "put": {
        "operationId": "put_project_api_v1_projects__project_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectConfig-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectConfig-Output"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Project"
      }
    },
    "/api/v1/projects/{project_id}/checkpoint": {
      "delete": {
        "operationId": "reset_project_checkpoint_api_v1_projects__project_id__checkpoint_delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckpointSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Project Checkpoint"
      },
      "get": {
        "operationId": "get_project_checkpoint_api_v1_projects__project_id__checkpoint_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckpointSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project Checkpoint"
      }
    },
    "/api/v1/projects/{project_id}/export": {
      "get": {
        "operationId": "export_project_api_v1_projects__project_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "json",
              "pattern": "^(json|yaml)$",
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Project"
      }
    },
    "/api/v1/projects/{project_id}/runs": {
      "post": {
        "operationId": "create_run_api_v1_projects__project_id__runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Run"
      }
    },
    "/api/v1/projects/{project_id}/validate": {
      "post": {
        "operationId": "validate_stored_project_api_v1_projects__project_id__validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Stored Project"
      }
    },
    "/api/v1/runs/{run_id}": {
      "delete": {
        "operationId": "cancel_run_api_v1_runs__run_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Run"
      },
      "get": {
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run"
      }
    },
    "/api/v1/runs/{run_id}/events": {
      "get": {
        "operationId": "get_run_events_api_v1_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Events"
      }
    },
...
  }
}